chore: Moves OnPasteSelect tests to component's folder
