+++ superset-frontend/src/components/Select/index.ts
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-export * from './SupersetStyledSelect';
+export * from './Select';
 export * from './styles';
-export { default } from './SupersetStyledSelect';
+export { default } from './Select';
 export { default as OnPasteSelect } from './OnPasteSelect';
